Move JWT authentication to its own file
