+++ DotNet8Auth/SqlScripts/DeleteAllFromLogs.sql
@@ -1,0 +1,1 @@
+Delete FROM [DotNet8AuthDb].[dbo].[Logs]
